+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.2.8 Minor improvements
+- Improvements in tomography object
+- Improvements in documentation (added return types hints)
+
 # 1.2.7 Bug solve
 - Solves #86: issues when loading a discrete number of frames from pco, png or tiff formats